// frontend/src/app/api/model/dishDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RecipeDto } from './recipeDto';
import { IngredientDto } from './ingredientDto';


export interface DishDto { 
    id?: string;
    name: string;
    preparedAt: string;
    calories?: number;
    protein?: number;
    fat?: number;
    carbohydrates?: number;
    recipes?: Array<RecipeDto>;
    ingredients?: Array<IngredientDto>;
}